Rename aura API to v6
