--- packages/graphql/aura-test-filter.js
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-"use strict";
-
-module.exports = function (tests) {
-    return {
-        filtered: tests
-            .filter((t) => {
-                return (
-                    t.includes("schema/aura-api") || t.includes("integration/aura-api") || t.includes("tck/aura-api")
-                );
-            })
-            .map((test) => ({ test })),
-    };
-};
