+++ packages/graphql/src/api-v6/graphQLTypeNames/EntityTypeNames.ts
