+++ packages/graphql/src/api-v6/graphQLTypeNames/TopLevelEntityTypeNames.ts
