+++ packages/graphql/src/api-v6/queryASTFactory/resolve-tree-parser/graphql-tree.ts
