+++ packages/graphql/src/api-v6/resolvers/mappers/connection-operation-mapper.ts
