+++ packages/graphql/src/api-v6/schema/schema-types/StaticTypes.ts
