+++ packages/graphql/tests/api-v6/integration/relationship.int.test.ts
@@ -21,7 +21,7 @@ import type { UniqueType } from "../../utils/graphql-types";
 import { TestHelper } from "../../utils/tests-helper";
 
 describe("Aura-api simple", () => {
-    const testHelper = new TestHelper({ cdc: false, auraApi: true });
+    const testHelper = new TestHelper({ cdc: false, v6Api: true });
 
     let Movie: UniqueType;
     let Actor: UniqueType;
